requirements.txt
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
db-dtypes==1.1.1
decorator==5.1.1
executing==1.2.0
fastavro==1.7.4
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.0
google-cloud-bigquery-storage==2.20.0
google-cloud-core==2.3.2
google-cloud-monitoring==2.15.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pandas-gbq==0.19.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
